Fix wrong type in printf(), introduced in revision 4265
git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@4278 c046a42c-6fe2-441c-8c8c-71466251a162

+++ vl.c
@@ -9000,8 +9000,8 @@ int main(int argc, char **argv)
     if (machine->ram_require & RAMSIZE_FIXED) {
         if (ram_size > 0) {
             if (ram_size < phys_ram_size) {
-                fprintf(stderr, "Machine `%s' requires %i bytes of memory\n",
-                                machine->name, phys_ram_size);
+                fprintf(stderr, "Machine `%s' requires %llu bytes of memory\n",
+                                machine->name, (unsigned long long) phys_ram_size);
                 exit(-1);
             }
